# api.yaml
openapi: 3.0.4
info:
  title: IotParking API
  description: REST API for the IotParking platform. Manages users, license plates, payments, preferences, parkings, active parking sessions, and bookings.
  version: 0.1.9

servers:
  - url: http://10.0.2.2:8000/v1
    description: Local development server

paths:

  # ── Authentication & Registration ──────────────────────────────────────────

  /login:
    post:
      summary: Authenticate user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, pass]
              properties:
                email:
                  type: string
                  format: email
                pass:
                  type: string
      responses:
        "200":
          description: Authenticated user details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  /register:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, pass]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                pass:
                  type: string
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "409":
          $ref: '#/components/responses/AlreadyExists'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── User ───────────────────────────────────────────────────────────────────

  /user:
    get:
      summary: Get user details
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id]
              properties:
                account_id:
                  type: integer
                name:
                  type: string
                email:
                  type: string
                  format: email
                new_pass:
                  type: string
                old_pass:
                  type: string
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    delete:
      summary: Delete user
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Plates ─────────────────────────────────────────────────────────────────

  /user/plates:
    get:
      summary: Get user's plates
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of user's plates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plate'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    post:
      summary: Add a new plate for the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, plate_text, plate_name, image_uri]
              properties:
                account_id:
                  type: integer
                plate_text:
                  type: string
                plate_name:
                  type: string
                image_uri:
                  type: string
      responses:
        "201":
          description: Plate created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plate'
        "409":
          $ref: '#/components/responses/AlreadyExists'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    delete:
      summary: Delete user's plate
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
        - name: plate_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plate deleted
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Payment ────────────────────────────────────────────────────────────────

  /user/payment:
    get:
      summary: Get user's payment method
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's payment method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update user's payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, payment]
              properties:
                account_id:
                  type: integer
                payment:
                  $ref: '#/components/schemas/Payment'
      responses:
        "200":
          description: Updated payment method
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Payment'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Preferences ────────────────────────────────────────────────────────────

  /user/preferences/distance:
    get:
      summary: Get user's distance preference
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's preferred maximum distance
          content:
            application/json:
              schema:
                type: object
                required: [distance_value]
                properties:
                  distance_value:
                    type: number
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update user's distance preference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, new_distance]
              properties:
                account_id:
                  type: integer
                new_distance:
                  type: number
      responses:
        "200":
          description: Updated preferred distance
          content:
            application/json:
              schema:
                type: object
                required: [distance_value]
                properties:
                  distance_value:
                    type: number
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  /user/preferences/price:
    get:
      summary: Get user's price preference
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User's preferred maximum price
          content:
            application/json:
              schema:
                type: object
                required: [price_value]
                properties:
                  price_value:
                    type: number
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update user's price preference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, new_price]
              properties:
                account_id:
                  type: integer
                new_price:
                  type: number
      responses:
        "200":
          description: Updated preferred price
          content:
            application/json:
              schema:
                type: object
                required: [price_value]
                properties:
                  price_value:
                    type: number
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Active parking sessions ────────────────────────────────────────────────

  /paying:
    get:
      summary: Get user's currently active parking sessions
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of user's currently active parking sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Session'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Parkings ───────────────────────────────────────────────────────────────

  /parkings/all:
    get:
      summary: Get a summary list of all parkings
      responses:
        "200":
          description: List of all parkings (summary)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    parking_id:
                      type: integer
                    parking_name:
                      type: string
                    address:
                      type: string

  /parkings:
    get:
      summary: Get parking details
      parameters:
        - name: parking_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parking details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    post:
      summary: Create a new parking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parking]
              properties:
                parking:
                  $ref: "#/components/schemas/Parking"
      responses:
        "201":
          description: Parking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        "409":
          $ref: '#/components/responses/AlreadyExists'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update parking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [parking]
              properties:
                parking:
                  $ref: "#/components/schemas/Parking"
      responses:
        "200":
          description: Parking updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    delete:
      summary: Delete parking
      parameters:
        - name: parking_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Parking deleted
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  /parkings/range:
    get:
      summary: Get parkings within a given range from a position
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lon
          in: query
          required: true
          schema:
            type: number
        - name: range
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of parkings within the specified range, with distances
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    parking:
                      $ref: '#/components/schemas/Parking'
                    distance:
                      type: number
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

  # ── Bookings ───────────────────────────────────────────────────────────────

  /bookings:
    get:
      summary: Get user's bookings
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of user's bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Booking'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    post:
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, booking]
              properties:
                account_id:
                  type: integer
                booking:
                  $ref: '#/components/schemas/Booking'
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        "409":
          $ref: '#/components/responses/AlreadyExists'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    patch:
      summary: Update user's booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, booking]
              properties:
                account_id:
                  type: integer
                booking:
                  $ref: '#/components/schemas/Booking'
      responses:
        "200":
          description: Updated booking detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Booking'
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

    delete:
      summary: Delete user's booking
      parameters:
        - name: account_id
          in: query
          required: true
          schema:
            type: integer
        - name: booking_id
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking deleted
        "422":
          $ref: '#/components/responses/ValidationErrorResponse'

### components ###

components:
  schemas:

    User:
      type: object
      required:
        - account_id
        - name
        - email
      properties:
        account_id:
          type: integer
          description: "Unique account identifier"
        name:
          type: string
          description: "Full name of the user"
        email:
          type: string
          format: email
          description: "Email address of the user"

    Plate:
      type: object
      required:
        - plate_id
        - plate_text
        - plate_name
        - is_active
        - image_uri
      properties:
        plate_id:
          type: integer
          description: "Unique identifier of the plate"
        plate_text:
          type: string
          description: "License plate number (e.g. AB123CD)"
        plate_name:
          type: string
          description: "Friendly name for the vehicle associated with the plate"
        is_active:
          type: boolean
          description: "Whether this plate is currently active"
        image_uri:
          type: string
          description: "URI of the plate image"

    Payment:
      type: object
      required:
        - circuit
        - card_number
      properties:
        circuit:
          type: string
          description: "Payment circuit (e.g. Visa, Mastercard)"
        card_number:
          type: string
          description: "Card number"

    Parking:
      type: object
      required:
        - parking_id
        - parking_name
        - total_slot
        - available_slot
        - price_per_hour
        - lat
        - lon
        - address
      properties:
        parking_id:
          type: integer
          description: "Unique identifier of the parking"
        parking_name:
          type: string
          description: "Name of the parking"
        total_slot:
          type: integer
          description: "Total number of parking slots"
        available_slot:
          type: integer
          description: "Number of currently available slots"
        price_per_hour:
          type: number
          description: "Hourly rate of the parking"
        lat:
          type: number
          description: "Latitude coordinate"
        lon:
          type: number
          description: "Longitude coordinate"
        address:
          type: string
          description: "Street address of the parking"

    Session:
      type: object
      required:
        - plate
        - parking
        - entry_time
        - amount
        - is_paid
      properties:
        plate:
          $ref: '#/components/schemas/Plate'
          description: "Details of the parked vehicle's plate"
        parking:
          $ref: '#/components/schemas/Parking'
          description: "Details of the parking facility"
        entry_time:
          type: string
          format: date-time
          description: "Date and time the vehicle entered the parking"
        amount:
          type: number
          description: "Amount due since entry"
        is_paid:
          type: boolean
          description: "Whether the session has been paid"

    Booking:
      type: object
      required:
        - booking_id
        - booking_name
        - parking
        - plate
        - date
        - days
        - slot_code
      properties:
        booking_id:
          type: integer
          description: "Unique identifier of the booking"
        booking_name:
          type: string
          description: "Friendly name of the booking"
        parking:
          $ref: '#/components/schemas/Parking'
          description: "Selected parking facility"
        plate:
          $ref: '#/components/schemas/Plate'
          description: "Selected plate for the booking"
        date:
          type: string
          format: date
          description: "Start date of the booking"
        days:
          type: integer
          description: "Duration of the booking in days"
        slot_code:
          type: integer
          description: "Code of the reserved parking slot"

    ValidationError:
      type: object
      required:
        - loc
        - msg
        - type
      properties:
        loc:
          description: "Location of the error (e.g. body, query, or array index)"
          type: array
          items:
            oneOf:
              - type: string
              - type: integer
        msg:
          type: string
          description: "Human-readable error message"
        type:
          type: string
          description: "Error type code"
        input:
          nullable: true
          description: "The value that caused the error"
        ctx:
          type: object
          description: "Additional context for the error"

    ConflictError:
      type: object
      required:
        - status
        - message
      properties:
        status:
          type: integer
          example: 409
        message:
          type: string
        field:
          type: string
          description: "The field that caused the conflict"
          example: "email"
  
    HTTPValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'

  responses:
    ValidationErrorResponse:
      description: "Validation Error (422)"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'

    AlreadyExists:
      description: "Cannot perform the action: resource already exists"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ConflictError'